+++ components/player/viewer.vue
@@ -21,7 +21,7 @@
     
     import brand from '~/brand/config'
 
-    import Button from '~/components/button'
+    import Button from '~/components/Button'
 
     export default {
         computed: {
