+++ pages/index.vue
@@ -25,7 +25,7 @@
     
     import brand from '~/brand/config'
 
-    import Button from '~/components/button'
+    import Button from '~/components/Button'
 
     export default {
         async asyncData(context) {
