+++ pages/room/create.vue
@@ -4,7 +4,7 @@
     </div>
 </template>
 <script>
-    import CreateRoom from '~/components/room/create'
+    import CreateRoom from '~/components/Room/Create'
 
     export default {
         head() {
